+++ src/components/IssueItem.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState, useContext } from "react";
 import { IDBIssue } from "../types/dbissues";
 import { setIssueTimeDB } from '../lib/localstore'
-import { CSSProperties } from "react";
 import { SettingsContext } from "../lib/SettingsContext";
+import { labelColor } from '../lib/componentHelpers'
 
 const editImg = require('../lib/img/Pencil-icon16.png')
 const saveImg = require('../lib/img/Save-icon16.png')
@@ -79,32 +79,6 @@ const IssueItem: React.FC<IDBIssue> = (props) => {
     const sTime = parseTime(time)
 
     return sTime.slice(0, -3)
-  }
-
-  const labelColor = (label_color: string): CSSProperties => {
-    const bgcolor: string = '#' + label_color || '#FFF'
-
-    const result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(bgcolor)
-    if (!result) {
-      return {}
-    }
-
-    const red = parseInt(result[1], 16)
-    const green = parseInt(result[2], 16)
-    const blue = parseInt(result[3], 16)
-
-    let color = ''
-    if (0.299 * red + 0.587 * green + 0.114 * blue > 127.5) {
-      color = '#020202'
-    } else {
-      color = '#FFF'
-    }
-
-    return {
-      color: color,
-      backgroundColor: bgcolor,
-
-    }
   }
 
   const editTimeHelper = () => {
